pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dashscope>=1.25.8",
    "jinja2>=3.1.6",
    "langchain>=1.2.6",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]
